requirements.txt
# Core UI + Voice
streamlit
torch
whisper
sounddevice
scipy
coqui-tts

# Document parsing
unstructured[pdf]
pypdf
python-docx
python-multipart


# Vector DB
chromadb

# LangChain ecosystem (v0.2+ only)
langchain>=0.2.0
langchain-community
langchain-ollama
langchain-chroma
streamlit-webrtc